requirements.txt
alembic==1.13.1
anyio==4.3.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.88.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.5
pydantic==1.10.14
PyJWT==2.6.0
sniffio==1.3.1
SQLAlchemy==2.0.0
starlette==0.22.0
typing_extensions==4.10.0
uvicorn==0.20.0
python-jose==3.3.0
passlib[bcrypt]
python-decouple
python-multipart